Moved another function from h3 to common
